+++ ErsatzTV.Application/ProgramSchedules/Commands/IProgramScheduleItemRequest.cs
@@ -16,11 +16,6 @@ namespace ErsatzTV.Application.ProgramSchedules.Commands
         int? MultipleCount { get; }
         TimeSpan? PlayoutDuration { get; }
         TailMode TailMode { get; }
-        ProgramScheduleItemCollectionType TailCollectionType { get; }
-        int? TailCollectionId { get; }
-        int? TailMultiCollectionId { get; }
-        int? TailSmartCollectionId { get; }
-        int? TailMediaItemId { get; }
         string CustomTitle { get; }
         GuideMode GuideMode { get; }
         int? PreRollFillerId { get; }